requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
anyio==3.7.1
async-timeout==3.0.1
attrs==23.1.0
black==23.7.0
black-pre-commit-mirror @ file:///home/dwayne/Devel/blood-on-the-clocktower
certifi==2023.7.22
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
discord.py==1.7.3
discord.py-stubs==1.7.3
distlib==0.3.7
exceptiongroup==1.1.2
Faker==19.2.0
filelock==3.12.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.5.26
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mypy==1.4.0
mypy-extensions==1.0.0
nodeenv==1.8.0
nodejs-bin==18.4.0a4
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.3.3
prisma==0.9.1
pydantic==1.10.12
pyright==1.1.320
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
virtualenv==20.24.2
yarl==1.9.2
